+++ src/main/java/com/bovo/Bovo/modules/kakaoLogin/kakao_dto/request/AuthorizationCodeDto.java
@@ -1,4 +1,4 @@
-package com.bovo.Bovo.modules.user.kakaoLogin.kakao_dto.request;
+package com.bovo.Bovo.modules.kakaoLogin.kakao_dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
