+++ src/main/java/com/bovo/Bovo/modules/kakaoLogin/repository/KakaoUserAuthRepositoryImpl.java
@@ -1,8 +1,8 @@
-package com.bovo.Bovo.modules.user.kakaoLogin.repository;
+package com.bovo.Bovo.modules.kakaoLogin.repository;
 
 import com.bovo.Bovo.common.User_Auth;
 import com.bovo.Bovo.common.Users;
-import com.bovo.Bovo.modules.user.kakaoLogin.kakao_dto.request.NewKakaoUserDto;
+import com.bovo.Bovo.modules.kakaoLogin.kakao_dto.request.NewKakaoUserDto;
 import jakarta.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
